# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API para gestionar usuarios, wishlists y libros",
        "title": "Wishlist API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
        "/users": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Lista los usuarios registrados",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/domain.User"
                            }
                        }
                    }
                }
            }
        },
        "/users/register": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Registrar nuevo usuario",
                "parameters": [
                    {
                        "description": "Usuario",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.RegisterUserRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/wishlist": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "wishlist"
                ],
                "summary": "Lista las wishlists del usuario",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "object",
                                "additionalProperties": true
                            }
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "wishlist"
                ],
                "summary": "Crea una nueva wishlist",
                "parameters": [
                    {
                        "description": "Datos de la lista",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.CreateWishlistRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
        "/wishlist/{id}": {
            "delete": {
                "tags": [
                    "wishlist"
                ],
                "summary": "Elimina una wishlist por ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID de la wishlist",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/wishlist/{id}/books": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "books"
                ],
                "summary": "Lista los libros de una wishlist",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID de la wishlist",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "object",
                                "additionalProperties": true
                            }
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "books"
                ],
                "summary": "Agrega un libro a la wishlist",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID de la wishlist",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Datos del libro",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.AddBookRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/wishlist/{id}/books/{bookID}": {
            "delete": {
                "tags": [
                    "books"
                ],
                "summary": "Elimina un libro de una wishlist",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID de la wishlist",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "ID del libro",
                        "name": "bookID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        }
    },
    "definitions": {
        "domain.User": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "password": {
                    "description": "hash",
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "handler.AddBookRequest": {
            "type": "object",
            "properties": {
                "author": {
                    "type": "string",
                    "example": "Antoine de Saint-Exupéry"
                },
                "title": {
                    "type": "string",
                    "example": "El Principito"
                }
            }
        },
        "handler.CreateWishlistRequest": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string",
                    "example": "Mi lista de libros"
                }
            }
        },
        "handler.RegisterUserRequest": {
            "type": "object",
            "properties": {
                "password": {
                    "type": "string",
                    "example": "1234"
                },
                "username": {
                    "type": "string",
                    "example": "david"
                }
            }
        }
    }
}